requirements.txt
flask==1.0.2
PyYAML>=3.13
psycopg2-binary
